feat: modularizing app in order to be accessed and mocked by the pytest coverage feature
