+++ app/app.py
@@ -3,8 +3,12 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_marshmallow import Marshmallow
 from dotenv import dotenv_values
 from flask_cors import CORS
+import os
 
-env_var = dotenv_values(".env")
+base_dir = os.path.abspath(os.path.dirname(__file__))
+env_path = os.path.join(base_dir, "..", ".env")
+
+env_var = dotenv_values(env_path)
 
 if env_var['ENVIRONMENT'] == 'testing':
     connection_uri = env_var['DB_URL_TESTING']
@@ -17,6 +21,9 @@ CORS(app)
 database = SQLAlchemy(app)
 ma = Marshmallow(app)
 
-from routes.users import *
+from app.routes.users import *
 
-from routes.tasks import *
+from app.routes.tasks import *
+
+if __name__ == "__main__":
+    app.run()
